Renaming file to follow redundant project name prefixes
